newsbot/requirements.txt
requests
python-dotenv
feedparser
google-generativeai
beautifulsoup4
lxml
telethon
tenacity
schedule
